+++ patches/Accessibility/strings_pl.cpp
@@ -117,6 +117,7 @@ const char* Get(Id id) {
         case Id::MenuCatMedical:       return "Medycyna";
         case Id::MenuCatMisc:          return "R\xF3\xBFne";
         case Id::MenuCatExplosives:    return "Materia\xB3y wybuchowe";
+        case Id::MenuCatCombatBehaviour: return "Zachowanie w walce";
         case Id::FmtMenuCatMulti:      return "%s: %s, %d opcji";
         case Id::FmtMenuCatSingle:     return "%s: %s";
         case Id::FmtMenuPlainMulti:    return "%s, %d opcji";
@@ -560,6 +561,7 @@ const char* Get(Id id) {
 
         case Id::FmtHelpNumberActions:   return "1 do 7: u\xBFycie ostatniej akcji z kategorii. 1 %s, 2 %s, 3 %s, 4 %s, 5 %s, 6 %s, 7 %s";
         case Id::HelpKeyOpenCategory:    return "Shift i 1 do 7: otwiera kategori\xEA, aby wybra\xE6 akcj\xEA";
+        case Id::FmtHelpNumberActions8:  return "8: zastosuj bie\xBF\xB9""ce %s. Shift i 8 otwiera kategori\xEA, aby wybra\xE6 inne";
         case Id::HelpKeyActionQueue:     return "Shift i H: otwiera kolejk\xEA akcji";
         case Id::HelpKeyLevelUp:         return "Shift i L: otwiera ekran awansu";
         case Id::HelpKeyCancelCombat:    return "F: przerywa walk\xEA";
@@ -641,10 +643,12 @@ const char* Get(Id id) {
         case Id::KbNamePersonalKey2:        return "Akcja osobista 2";
         case Id::KbNamePersonalKey3:        return "Akcja osobista 3";
         case Id::KbNamePersonalKey4:        return "Akcja osobista 4";
+        case Id::KbNamePersonalKey5:        return "Akcja osobista 5";
         case Id::KbNameActionBarOpen1:      return "Otw\xF3rz pasek akcji 1";
         case Id::KbNameActionBarOpen2:      return "Otw\xF3rz pasek akcji 2";
         case Id::KbNameActionBarOpen3:      return "Otw\xF3rz pasek akcji 3";
         case Id::KbNameActionBarOpen4:      return "Otw\xF3rz pasek akcji 4";
+        case Id::KbNameActionBarOpen5:      return "Otw\xF3rz pasek akcji 5";
         case Id::KbNameTargetActionOpen1:   return "Otw\xF3rz akcj\xEA celu 1";
         case Id::KbNameTargetActionOpen2:   return "Otw\xF3rz akcj\xEA celu 2";
         case Id::KbNameTargetActionOpen3:   return "Otw\xF3rz akcj\xEA celu 3";